First deployable commit.
Implemented the most fundamental multiprocessing utility functions such
as functions responsible for the handling of the thread pool, within
which all updates are supposed to be handled. Also implemented the
UpdateReceiver and MessageUtils classes and the UpnotifyBot class. Most
implementations had already been commented and rest will also be
commented with javadoc later on.